changed qc profiling to use doubles instead of ints for profile counters, so it can now count beyond 4 billion
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6379 d7cf8633-e32d-0410-b094-e92efae38249

+++ pr_comp.h
@@ -163,9 +163,10 @@ typedef struct mfunction_s
 	int		parm_start;
 	int		locals;				// total ints of parms + locals
 
-	int		profile;		// runtime
-	int		builtinsprofile; // cost of builtin functions called by this function
-	int		callcount; // times the functions has been called since the last profile call
+	// these are doubles so that they can count up to 54bits or so rather than 32bit
+	double	profile;		// runtime
+	double	builtinsprofile; // cost of builtin functions called by this function
+	double	callcount; // times the functions has been called since the last profile call
 
 	int		s_name;
 	int		s_file;			// source file defined in
